split out some intake tests into separate files
